Bugs Fixed

+++ routes/web.php
@@ -46,21 +46,21 @@ Route::middleware(['auth', 'admin.auth'])
         Route::put('pages/{page}', [PageController::class, 'update'])->middleware('permission:pages.update')->name('pages.update');
         Route::delete('pages/{page}', [PageController::class, 'destroy'])->middleware('permission:pages.delete')->name('pages.destroy');
 
-        Route::get('headers', [HeaderTemplateController::class, 'index'])->name('headers.index');
-        Route::get('headers/create', [HeaderTemplateController::class, 'create'])->name('headers.create');
-        Route::post('headers', [HeaderTemplateController::class, 'store'])->name('headers.store');
-        Route::get('headers/{headerTemplate}/edit', [HeaderTemplateController::class, 'edit'])->name('headers.edit');
-        Route::put('headers/{headerTemplate}', [HeaderTemplateController::class, 'update'])->name('headers.update');
-        Route::delete('headers/{headerTemplate}', [HeaderTemplateController::class, 'destroy'])->name('headers.destroy');
-        Route::post('headers/{headerTemplate}/set-default', [HeaderTemplateController::class, 'setDefault'])->name('headers.set-default');
+        Route::get('headers', [HeaderTemplateController::class, 'index'])->middleware('permission:headers.manage')->name('headers.index');
+        Route::get('headers/create', [HeaderTemplateController::class, 'create'])->middleware('permission:headers.manage')->name('headers.create');
+        Route::post('headers', [HeaderTemplateController::class, 'store'])->middleware('permission:headers.manage')->name('headers.store');
+        Route::get('headers/{headerTemplate}/edit', [HeaderTemplateController::class, 'edit'])->middleware('permission:headers.manage')->name('headers.edit');
+        Route::put('headers/{headerTemplate}', [HeaderTemplateController::class, 'update'])->middleware('permission:headers.manage')->name('headers.update');
+        Route::delete('headers/{headerTemplate}', [HeaderTemplateController::class, 'destroy'])->middleware('permission:headers.manage')->name('headers.destroy');
+        Route::post('headers/{headerTemplate}/set-default', [HeaderTemplateController::class, 'setDefault'])->middleware('permission:headers.manage')->name('headers.set-default');
 
-        Route::get('footers', [FooterTemplateController::class, 'index'])->name('footers.index');
-        Route::get('footers/create', [FooterTemplateController::class, 'create'])->name('footers.create');
-        Route::post('footers', [FooterTemplateController::class, 'store'])->name('footers.store');
-        Route::get('footers/{footerTemplate}/edit', [FooterTemplateController::class, 'edit'])->name('footers.edit');
-        Route::put('footers/{footerTemplate}', [FooterTemplateController::class, 'update'])->name('footers.update');
-        Route::delete('footers/{footerTemplate}', [FooterTemplateController::class, 'destroy'])->name('footers.destroy');
-        Route::post('footers/{footerTemplate}/set-default', [FooterTemplateController::class, 'setDefault'])->name('footers.set-default');
+        Route::get('footers', [FooterTemplateController::class, 'index'])->middleware('permission:footers.manage')->name('footers.index');
+        Route::get('footers/create', [FooterTemplateController::class, 'create'])->middleware('permission:footers.manage')->name('footers.create');
+        Route::post('footers', [FooterTemplateController::class, 'store'])->middleware('permission:footers.manage')->name('footers.store');
+        Route::get('footers/{footerTemplate}/edit', [FooterTemplateController::class, 'edit'])->middleware('permission:footers.manage')->name('footers.edit');
+        Route::put('footers/{footerTemplate}', [FooterTemplateController::class, 'update'])->middleware('permission:footers.manage')->name('footers.update');
+        Route::delete('footers/{footerTemplate}', [FooterTemplateController::class, 'destroy'])->middleware('permission:footers.manage')->name('footers.destroy');
+        Route::post('footers/{footerTemplate}/set-default', [FooterTemplateController::class, 'setDefault'])->middleware('permission:footers.manage')->name('footers.set-default');
 
         Route::get('media', [MediaController::class, 'index'])->middleware('permission:media.view')->name('media.index');
         Route::delete('media/{medium}', [MediaController::class, 'destroy'])->middleware('permission:media.delete')->name('media.destroy');
@@ -96,13 +96,13 @@ Route::middleware(['auth', 'admin.auth'])
         // ============================================================
         // ========== MENU ROUTES (PHASE 2) ==========================
         // ============================================================
-        Route::get('/menus', [MenuController::class, 'index'])->name('menus.index');
-        Route::get('/menus/create', [MenuController::class, 'create'])->name('menus.create');
-        Route::post('/menus', [MenuController::class, 'store'])->name('menus.store');
-        Route::get('/menus/{menu}/edit', [MenuController::class, 'edit'])->name('menus.edit');
-        Route::put('/menus/{menu}', [MenuController::class, 'update'])->name('menus.update');
-        Route::delete('/menus/{menu}', [MenuController::class, 'destroy'])->name('menus.destroy');
-        Route::post('/menus/{menu}/set-default', [MenuController::class, 'setDefault'])->name('menus.set-default');
+        Route::get('/menus', [MenuController::class, 'index'])->middleware('permission:menus.manage')->name('menus.index');
+        Route::get('/menus/create', [MenuController::class, 'create'])->middleware('permission:menus.manage')->name('menus.create');
+        Route::post('/menus', [MenuController::class, 'store'])->middleware('permission:menus.manage')->name('menus.store');
+        Route::get('/menus/{menu}/edit', [MenuController::class, 'edit'])->middleware('permission:menus.manage')->name('menus.edit');
+        Route::put('/menus/{menu}', [MenuController::class, 'update'])->middleware('permission:menus.manage')->name('menus.update');
+        Route::delete('/menus/{menu}', [MenuController::class, 'destroy'])->middleware('permission:menus.manage')->name('menus.destroy');
+        Route::post('/menus/{menu}/set-default', [MenuController::class, 'setDefault'])->middleware('permission:menus.manage')->name('menus.set-default');
 
         // ============================================================
         // ========== PAGE VERSIONING ROUTES (PHASE 2) ================
@@ -130,9 +130,7 @@ Route::middleware('throttle:60,1')->group(function () {
     Route::get('/pages/{slug}', [PublicController::class, 'showPage'])->name('public.page');
 });
 
-Route::get('/dashboard', function () {
-    return view('dashboard');
-})->middleware(['auth', 'verified'])->name('dashboard');
+
 
 Route::middleware(['auth', 'activity'])->group(function () {
     Route::get('/profile', [ProfileController::class, 'edit'])->name('profile.edit');
